ISW_2025-01/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
azure-core==1.34.0
azure-storage-blob==12.25.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==45.0.2
cssselect==1.2.0
Django==5.1.2
djangorestframework==3.15.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.0
h5py==3.12.1
huggingface-hub==0.26.2
idna==3.10
isodate==0.7.2
jieba3k==0.35.1
joblib==1.4.2
keras==3.6.0
libclang==18.1.1
lxml==5.3.0
lxml_html_clean==0.4.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
protobuf==5.28.3
psycopg==3.2.3
pycparser==2.22
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
safetensors==0.4.5
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.20.3
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
